+++ src/state/index.ts
@@ -3,4 +3,3 @@
  */
 
 export { useThemeStore } from './themeStore';
-export { useDevLogStore, selectFilteredLogs, LOG_MODULES } from './devLogStore';